Functioning PyInstrument
