+++ pyproject.toml
@@ -61,6 +61,7 @@ channels = {version = ">=3.0.5", optional = true}
 astunparse = {version = "^1.6.3", python = "<3.9"}
 libcst = {version = ">=0.4.7", optional = true}
 rich = {version = ">=12.0.0", optional = true}
+pyinstrument = "^4.4.0"
 
 
 [tool.poetry.group.dev.dependencies]
