+++ strawberry/extensions/pyinstrument.py
@@ -1,11 +1,9 @@
-from typing import Iterator, List, Type
-
-from strawberry.extensions.base_extension import SchemaExtension
+from typing import Iterator, Union, TextIO
+from pathlib import Path
 
 from pyinstrument import Profiler
 
-# if TYPE_CHECKING:
-#    from graphql import ASTValidationRule
+from strawberry.extensions.base_extension import SchemaExtension
 
 
 class PyInstrument(SchemaExtension):
@@ -13,7 +11,7 @@ class PyInstrument(SchemaExtension):
     Extension to profile the execution time of resolvers using PyInstrument.
     """
 
-    def __init__(self, report_path: str):
+    def __init__(self, report_path: Path):
         self.report_path = report_path
 
     def on_operation(self) -> Iterator[None]:
@@ -22,6 +20,9 @@ class PyInstrument(SchemaExtension):
         in this case we start the profiler and yield
         then we stop the profiler when the operation is done
         """
+
+        # self.execution_context
+
         # Start the profiler
         profiler = Profiler()
         profiler.start()
@@ -32,4 +33,4 @@ class PyInstrument(SchemaExtension):
         # Stop the profiler
         profiler.stop()
         with open(self.report_path, "w", encoding="utf-8") as f:
-            profiler.output_html(f)
+            f.write(profiler.output_html())
